ommitted unnecessary files and joined testsets in test.py
