--- sessions/panda.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-echo $1
-echo '\n'
-exec "$SHELL"
